+++ src/main/java/com/kupofficial/firstwebproject/CalculatorService.java
@@ -5,5 +5,5 @@ public interface CalculatorService {
     String sumOperation(int a, int b);
     String minusOperation(int a, int b);
     String multiplyOperation(int a, int b);
-    String divideOperation(int a, int b);
+    String divideOperation(double a, double b);
 }
